requirements.txt
langchain
langchain-google-genai
langchain-community
chromadb
tiktoken
GitPython
python-dotenv
streamlit
nest_asyncio